+++ PayPalNativePayments/src/main/java/com/paypal/android/nativepayments/PayPalNativeCheckoutError.kt
@@ -1,4 +1,4 @@
-package com.paypal.android.checkout
+package com.paypal.android.nativepayments
 
 import com.paypal.android.corepayments.PayPalSDKError
 import com.paypal.checkout.error.ErrorInfo
