+++ PayPalNativePayments/src/main/java/com/paypal/android/nativepayments/PayPalNativeCheckoutListener.kt
@@ -1,4 +1,4 @@
-package com.paypal.android.checkout
+package com.paypal.android.nativepayments
 
 import com.paypal.android.corepayments.PayPalSDKError
 import com.paypal.checkout.shipping.ShippingChangeActions
