+++ PaymentButtons/src/main/java/com/paypal/android/paymentbuttons/PayPalCreditButton.kt
@@ -4,7 +4,7 @@ import android.content.Context
 import android.content.res.TypedArray
 import android.util.AttributeSet
 import androidx.core.content.res.use
-import com.paypal.android.checkout.paymentbutton.error.createFormattedIllegalArgumentException
+import com.paypal.android.nativepayments.paymentbutton.error.createFormattedIllegalArgumentException
 import com.paypal.android.ui.R
 import com.paypal.android.paymentbuttons.PayPalCreditButtonColor.BLACK
 import com.paypal.android.paymentbuttons.PayPalCreditButtonColor.DARK_BLUE
